refactor(check): Move error types to their own module
